# Conflicts:
#	src/main/java/org/example/TripPlanner.java
#	src/main/java/org/example/repository/bookingRepo/BookingRepository.java
#	src/main/java/org/example/services/BookingService.java

+++ src/main/java/org/example/repository/userRepo/UserRepository.java
@@ -5,7 +5,7 @@ import org.example.entity.User;
 import java.util.List;
 
 public interface UserRepository {
-    User get(Long id);
+    User get(int id);
     void add(User user);
     void update(User user);
     void remove(User user);
